corrected compilation errors

+++ Server.hpp
@@ -6,7 +6,7 @@
 /*   By: cbouvet <cbouvet@student.42lisboa.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/02 20:46:31 by cbouvet           #+#    #+#             */
-/*   Updated: 2025/06/29 18:47:12 by cbouvet          ###   ########.fr       */
+/*   Updated: 2025/06/29 20:32:07 by cbouvet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -121,9 +121,9 @@ class Server
 		bool	authCheck(Client &client);
 		int		cmdCheck(Client *client, Channel *channel, std::string target);
 		int		checkModeFormat(str_vector const &msg);
-		void	sendMode(Client &client, Channel &channel, int stop, str_vector const &msg);
+		void	sendMode(Client &client, Channel &channel, size_t stop, str_vector const &msg);
 		std::string nameList(Channel &channel);
 		str_vector newVector(std::string const &arg1, std::string const &arg2, std::string *arg3);
-		std::string	unSplit(str_vector const &msg, int index);
+		std::string	unSplit(str_vector const &msg, size_t index);
 	};
